Awards 1.12.2 (#1349)
- Most changes from #1249
- Removed usage of ilch-date CSS class (#1232)
- Several fixes for stuff that broke.

+++ application/modules/awards/controllers/admin/Index.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @copyright Ilch 2
  * @package ilch
@@ -47,8 +48,7 @@ class Index extends \Ilch\Controller\Admin
             $items[0]['active'] = true;
         }
 
-        $this->getLayout()->addMenu
-        (
+        $this->getLayout()->addMenu(
             'menuAwards',
             $items
         );
@@ -76,7 +76,7 @@ class Index extends \Ilch\Controller\Admin
 
         $awards = $awardsMapper->getAwards();
         foreach ($awards as $award) {
-            foreach($award->getRecipients() as $recipient) {
+            foreach ($award->getRecipients() as $recipient) {
                 if ($recipient->getTyp() == 1) {
                     $userIds[] = $recipient->getUtId();
                 } else {
@@ -104,56 +104,46 @@ class Index extends \Ilch\Controller\Admin
         $awardsMapper = new AwardsMapper();
         $userMapper = new UserMapper();
 
+        $awardsModel = new AwardsModel();
         if ($this->getRequest()->getParam('id')) {
             $this->getLayout()->getAdminHmenu()
                 ->add($this->getTranslator()->trans('menuAwards'), ['action' => 'index'])
                 ->add($this->getTranslator()->trans('edit'), ['action' => 'treat']);
 
-            $awards = $awardsMapper->getAwardsById($this->getRequest()->getParam('id'));
+            $awardsModel = $awardsMapper->getAwardsById($this->getRequest()->getParam('id'));
 
-            if (!$awards) {
+            if (!$awardsModel) {
                 $this->redirect()
                     ->withMessage('awardNotFound', 'danger')
                     ->to(['action' => 'index']);
             }
-
-            $this->getView()->set('awards', $awards);
         } else {
             $this->getLayout()->getAdminHmenu()
                 ->add($this->getTranslator()->trans('menuAwards'), ['action' => 'index'])
                 ->add($this->getTranslator()->trans('add'), ['action' => 'treat']);
         }
+        $this->getView()->set('award', $awardsModel);
 
         if ($this->getRequest()->isPost()) {
-            // Add BASE_URL if image starts with application to get a complete URL for validation
-            $image = trim($this->getRequest()->getPost('image'));
-            if (!empty($image) && strncmp($image, 'application', 11) === 0) {
-                $image = BASE_URL.'/'.$image;
-            }
+            $post = $this->getRequest()->getPost();
 
-            $post = [
-                'date' => trim($this->getRequest()->getPost('date')),
-                'rank' => trim($this->getRequest()->getPost('rank')),
-                'image' => $image,
-                'utId' => $this->getRequest()->getPost('utId'),
-                'event' => trim($this->getRequest()->getPost('event')),
-                'page' => trim($this->getRequest()->getPost('page'))
-            ];
+            // Add BASE_URL if image starts with application to get a complete URL for validation
+            if (!empty($post['image']) && strncmp($post['image'], 'application', 11) === 0) {
+                $post['image'] = BASE_URL . '/' . $post['image'];
+            }
 
             Validation::setCustomFieldAliases([
                 'utId' => 'invalidUserTeam',
             ]);
 
             $validation = Validation::create($post, [
-                'date'  => 'required',
+                'date'  => 'required|date:d.m.Y',
                 'rank'  => 'required|numeric|integer|min:1',
                 'image' => 'url',
                 'utId'  => 'required',
                 'event' => 'required',
                 'page' => 'url'
             ]);
-
-            $post['image'] = trim($this->getRequest()->getPost('image'));
 
             if ($validation->isValid()) {
                 $recipientMapper = new RecipientsMapper();
@@ -162,15 +152,15 @@ class Index extends \Ilch\Controller\Admin
                 if ($this->getRequest()->getParam('id')) {
                     $awardsModel->setId($this->getRequest()->getParam('id'));
                 }
-                $awardsModel->setDate(new \Ilch\Date($post['date']))
-                    ->setRank($post['rank'])
-                    ->setImage($post['image'])
-                    ->setEvent($post['event'])
-                    ->setURL($post['page']);
+                $awardsModel->setDate(new \Ilch\Date($this->getRequest()->getPost('date')))
+                    ->setRank($this->getRequest()->getPost('rank', 0))
+                    ->setImage($this->getRequest()->getPost('image', ''))
+                    ->setEvent($this->getRequest()->getPost('event', ''))
+                    ->setURL($this->getRequest()->getPost('page', ''));
                 $idOrAffectedRows = $awardsMapper->save($awardsModel);
 
                 $recipientModels = [];
-                foreach($post['utId'] as $value) {
+                foreach ($post['utId'] as $value) {
                     $recipientModel = new RecipientModel();
                     $recipientModel->setAwardId((!$this->getRequest()->getParam('id')) ? $idOrAffectedRows : $awardsModel->getId())
                         ->setUtId(substr($value, 2))
@@ -184,7 +174,7 @@ class Index extends \Ilch\Controller\Admin
                     $notificationsMapper = new NotificationsMapper();
                     $notificationModels = [];
 
-                    foreach($recipientModels as $recipientModel) {
+                    foreach ($recipientModels as $recipientModel) {
                         $users = [];
                         if ($recipientModel->getTyp() === 2) {
                             // The recipient is a team.
@@ -242,7 +232,7 @@ class Index extends \Ilch\Controller\Admin
 
     public function delAction()
     {
-        if ($this->getRequest()->isSecure()) {
+        if ($this->getRequest()->isSecure() && !empty($this->getRequest()->getParam('id'))) {
             $awardsMapper = new AwardsMapper();
             $awardsMapper->delete($this->getRequest()->getParam('id'));
